+++ supercell/logging.py
@@ -18,7 +18,6 @@
 from __future__ import (absolute_import, division, print_function,
                         with_statement)
 
-from logging import Formatter
 from logging.handlers import TimedRotatingFileHandler
 import os
 
@@ -32,9 +31,3 @@ class SupercellLoggingHandler(TimedRotatingFileHandler):
         logfile = logfile % {'pid': os.getpid()}
         TimedRotatingFileHandler.__init__(self, logfile, when='d',
                                           interval=1, backupCount=10)
-        formatter = Formatter(' - '.join(['[%(asctime)s',
-                                          '%(relativeCreated).3f]',
-                                          '%(name)s',
-                                          '%(levelname)s',
-                                          '%(message)s']))
-        self.setFormatter(formatter)