+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 0.6.0 -
 ---------------------------
 
+- allow logging to `stdout`
 - Enable log file name with pid
 - General base class for middleware decorators
 - Typed query params deal with validation of query params